requirements.txt
django>=4.2,<5.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
drf-spectacular>=0.27,<1.0
django-cors-headers>=4.4,<5.0
django-ckeditor-5>=0.2,<1.0
django-unfold>=0.47,<1.0
dj-database-url>=2.2,<3.0
django-environ>=0.11,<1.0
gunicorn>=22.0,<24.0
psycopg2-binary>=2.9,<3.0
Pillow>=10.0,<12.0
whitenoise>=6.7,<7.0
pytest>=8.0,<9.0
pytest-django>=4.8,<5.0
ruff>=0.6,<1.0
